jobsearch_agents/requirements.txt
google-adk
praw
python-dotenv 
requests
fastapi
uvicorn
aiohttp
requests
firebase-admin
google-cloud-storage
deprecated